Added non-working formula for openbabel.
For me it fails on the cmake stage, but if I run it with debug mode:

    brew install -vd openbabel

Then when it fails I do precisely what it said it tried to do, it works.
I don't understand the problem. Doing it this way I can get it to continue
and install everything fine. Any suggestions?